+++ src/NewId/Formatters/HexFormatter.cs
@@ -10,7 +10,7 @@
             _alpha = upperCase ? 'A' : 'a';
         }
 
-        public string Format(int[] bytes)
+        public string Format(byte[] bytes)
         {
             var result = new char[32];
